pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
django = "3.2.11"
asgiref = "3.6.0"
cffi = "1.15.1"
cryptography = "39.0.0"
dajax = "1.3"
django-dajax = "0.9.2"
django-dajaxice = "0.7"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
jinja2 = "3.1.2"
jwt = "1.3.1"
markupsafe = "2.1.1"
pillow = "9.4.0"
pycparser = "2.21"
pyjwt = "2.6.0"
pytz = "2022.7"
sqlparse = "0.4.3"
tzdata = "2022.7"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"